requirements.txt
altair==5.0.1
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
black==23.7.0
blinker==1.6.2
bokeh==3.2.1
bs4==0.0.1
cachetools==5.3.1
catboost==1.2
category-encoders==2.6.1
causalgraphicalmodels==0.0.4
CausalInference==0.1.3
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.6
cloudpickle==2.2.1
cmake==3.26.4
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
filelock==3.12.2
fonttools==4.39.4
fsspec==2023.10.0
gdown==4.7.1
gitdb==4.0.10
GitPython==3.1.32
graphviz==0.20.1
idna==3.4
imageio==2.31.1
imbalanced-learn==0.11.0
imblearn==0.0
implicit==0.7.0
importlib-metadata==6.8.0
ipykernel==6.23.1
ipython==8.14.0
ipywidgets==8.0.7
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyterlab-widgets==3.0.8
kiwisolver==1.4.4
lazy_loader==0.3
lightgbm==4.0.0
lime==0.2.0.1
lit==16.0.6
llvmlite==0.40.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==3.1
numba==0.57.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
opencv-python==4.8.0.76
packaging==23.1
pandas==2.0.2
parso==0.8.3
pathspec==0.11.2
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.1
plotly==5.15.0
prompt-toolkit==3.0.38
protobuf==4.23.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pydeck==0.8.0
Pygments==2.15.1
Pympler==1.0.1
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.1.0
referencing==0.30.2
requests==2.31.0
rich==13.5.2
rpds-py==0.9.2
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
shap==0.42.1
six==1.16.0
slicer==0.0.7
smmap==5.0.0
soupsieve==2.4.1
stack-data==0.6.2
statsmodels==0.14.0
streamlit==1.25.0
sympy==1.12
tenacity==8.2.2
threadpoolctl==3.1.0
tifffile==2023.7.18
toml==0.10.2
toolz==0.12.0
torch==2.1.0
torchvision==0.16.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
triton==2.1.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.3
validators==0.20.0
watchdog==3.0.0
wcwidth==0.2.6
widgetsnbextension==4.0.8
xgboost==1.7.6
xyzservices==2023.7.0
zipp==3.16.2
